+++ noise.h
@@ -10,32 +10,32 @@
 // So may the world be empty, no beauty, no fun.
 class PerlinNoise {
 public:
-	static float get(float x, float y, float z);
+	static double get(double x, double y, double z);
 
 private:
-	class Vector : public std::tuple<float, float, float> {
+	class Vector : public std::tuple<double, double, double> {
 	public:
-		Vector(float x, float y, float z);
+		Vector(double x, double y, double z);
 
-		float x() const;
-		float y() const;
-		float z() const;
-		float dot(const Vector &v) const;
+		double x() const;
+		double y() const;
+		double z() const;
+		double dot(const Vector &v) const;
 		Vector sub(const Vector &v) const; // Overload? Well, if one would...
 		                                   // One will only choose what they think that they should
 	};
 
-	static Vector grad_vector(int x, int y, int z);
+	static Vector grad_vector(double x, double y, double z);
 	static Vector grad_vector_from_corner(const Vector &v);
 	static std::array<Vector, 8> cell_corners(const Vector &v);
 	static std::array<Vector, 8> offset_vectors(const Vector &v);
-	static std::array<float, 8> cell_dots(const Vector &v);
-	static float cell_interpolate(std::array<float, 8> dots, const Vector &v);
-	static float interpolate(float a, float b, float w);
+	static std::array<double, 8> cell_dots(const Vector &v);
+	static double cell_interpolate(std::array<double, 8> dots, const Vector &v);
+	static double interpolate(double a, double b, double w);
 };
 
 class Noise {
 public:
-	static float wiggle(float base, float min, float max, float step);
-	static float random();
+	static double wiggle(double base, double min, double max, double step);
+	static double random();
 };
